Add an on-disk cache for FlightInfos objects
A typical behaviour on my side is to take a look at the departures and
when there's something abnormal with a given aircraft, search for all
the flights (arrivals and departures) of the plane. For this I execute
another call to gvacli that I feed to grep to search for the
registration. This issues another unnecessary network call as the
information is very unlikely to change in such a short period of time.
Hence, adding a short-lived cache (60 seconds, currently hardcoded) is
good for both the user (faster output) and the server (less unnecessary
requests).

The patch adds a new command line option (--no-cache) to ignore the
local cache, even if it's still fresh. My intention was to never fail
hard if the cache cannot be generated/read/etc.

Multi OS compatibility is kept as os.UserCacheDir() is portable.