Reformat "Next Info" as "Delayed" (#11)
Departures with status "Next Info" are departures that are officially
delayed for which the expected departure time is unknown. This status
comes with an expected time for more info that's available on the
NextPublicAdvice field.

I don't believe that the "Delayed" status ever applies to departures, it
seems to me that delayed departures have either:

 * An empty status and an expected (delayed) departure time.
 * "Next info" status, no expected departure time and NextPublicAdvice.

Ideally if the status is "Next Info" we should be printing
NextPublicAdvice somehere but for the time being translating "Next Info"
into "Delayed" is enough.